requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0
stripe==7.8.0
razorpay==1.4.1
requests==2.31.0
supabase==2.3.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
supabase==2.4.0
